Updated lesson rates on home page.
I updated the lesson rates table on the home page to reflect the fact
that lessons do not end on a full hour, but rather have a 15 minute gap
for George to prepare for the next lesson.

I also added the _site directory to the .gitignore file so that it will
not show up in git status.